chore: ignore bridgecrewio/checkov-action
Update the config to ignore the bridgecrewio/checkov-action GitHub
Action.  The reason for this is that the bridgecrewio GitHub org has
an IP allow list that's currently blocking the lookup of this dependency
by the Renovate GitHub app's worker.

As a result, Renovate PRs and the dependency dashboard are showing
a warning.  An issue has been created with the action authors to
see if the IP allow list can be updated.