fix: update to new minimumReleaseAge setting
The `minimumReleaseAge` replaces `stabilityDays` and will cause Renovate to
mark any dependency upgrade that is less than 7 days old as pending.  This
setting is effective at blocking malicious package updates as they are usually
caught within a week of their release.

Additionaly the `minimumReleaseAgeBehaviour` will cause any dependency
update with a timestemp to be marked as pending.