fix: update to new minimumReleaseAge setting (#109)
The `minimumReleaseAge` replaces `stabilityDays` and will cause Renovate to
mark any dependency upgrade that is less than 7 days old as pending.  This
setting is effective at blocking malicious package updates as they are usually
caught within a week of their release.

Additionally the `minimumReleaseAgeBehaviour` will cause any dependency
update with a timestamp to be marked as pending.