fix: remove minimumReleaseAge check for lockfile maintenance (#113)
Renovate is unable to validate the release age during lockfile maintenance because
it delegates this behaviour to the package manager.  As such, all lockfile maintenance
PRs will shows as failing the `renovate/stability-days` check.

Where possible, this minimum release age check needs to be added to the package
manager's configuration.